# 资源/apifox-openapi.yaml
openapi: 3.0.3
info:
  title: Hope and Sparks API
  version: 1.0.0
  description: 'Apifox 导入版接口规范。详细业务说明、数据库落库规则和枚举映射见根目录

    `Hope_and_Sparks_API接口文档.md`。

    '
servers:
- url: http://localhost:8080
  description: 本地开发环境
- url: /
  description: 相对路径
security:
- bearerAuth: []
...
paths:
  /api/v1/auth/register:
    post:
      tags:
      - 用户与认证
      summary: 注册
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                email:
                  type: string
                  example: spark@example.com
                  description: 邮箱，对应 sys_user.email
                password:
                  type: string
                  example: '12345678'
                  description: 密码，服务端加密后写入 sys_user.password_hash
                nickname:
                  type: string
                  example: 一粒黑子
                  description: 昵称，对应 sys_user.nickname
                role:
                  type: string
                  example: spark
                  description: 用户角色，默认 spark
              required:
              - email
              - password
              - nickname
      responses:
        '200':
          $ref: '#/components/responses/Success'
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Body | `email` | 是 | string | 邮箱，对应 sys_user.email |

        | Body | `password` | 是 | string | 密码，服务端加密后写入 sys_user.password_hash |

        | Body | `nickname` | 是 | string | 昵称，对应 sys_user.nickname |

        | Body | `role` | 否 | string | 用户角色，默认 spark |


        主要数据表：sys_user、user_profile、user_login_session'
  /api/v1/auth/login:
    post:
      tags:
      - 用户与认证
      summary: 登录
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                account:
                  type: string
                  example: spark@example.com
                  description: 邮箱或用户名，对应 sys_user.email/sys_user.username
                password:
                  type: string
                  example: '12345678'
                  description: 密码，服务端加密后写入 sys_user.password_hash
                clientType:
                  type: string
                  example: web
                  enum:
                  - web
                  - desktop
                  - mobile
                  description: 客户端类型：web/desktop/mobile
                deviceId:
                  type: string
                  example: device_001
                  description: 设备标识，对应 user_login_session.device_id
              required:
              - account
              - password
      responses:
        '200':
          $ref: '#/components/responses/AuthSuccess'
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Body | `account` | 是 | string | 邮箱或用户名，对应 sys_user.email/sys_user.username |

        | Body | `password` | 是 | string | 密码，服务端加密后写入 sys_user.password_hash |

        | Body | `clientType` | 否 | string | 客户端类型：web/desktop/mobile |

        | Body | `deviceId` | 否 | string | 设备标识，对应 user_login_session.device_id |


        主要数据表：sys_user、user_profile、user_login_session'
  /api/v1/auth/refresh:
    post:
      tags:
      - 用户与认证
      summary: 刷新 token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                refreshToken:
                  type: string
                  description: 刷新令牌，对应 user_login_session.refresh_token
              required:
              - refreshToken
      responses:
        '200':
          $ref: '#/components/responses/AuthSuccess'
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Body | `refreshToken` | 是 | string | 刷新令牌，对应 user_login_session.refresh_token |


        主要数据表：sys_user、user_profile、user_login_session'
...
  /api/v1/auth/password/reset-request:
    post:
      tags:
      - 用户与认证
      summary: 找回密码
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                email:
                  type: string
                  description: 邮箱，对应 sys_user.email
                  format: email
                  example: spark@example.com
              required:
              - email
      responses:
        '200':
          $ref: '#/components/responses/Success'
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Body | `email` | 是 | string | 邮箱，对应 sys_user.email |


        主要数据表：sys_user、user_profile、user_login_session'
  /api/v1/auth/password/reset-confirm:
    post:
      tags:
      - 用户与认证
      summary: 重置密码
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                token:
                  type: string
                  description: 重置或校验凭证
                password:
                  type: string
                  description: 密码，服务端加密后写入 sys_user.password_hash
                  example: '12345678'
              required:
              - token
              - password
      responses:
        '200':
          $ref: '#/components/responses/Success'
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Body | `token` | 是 | string | 重置或校验凭证 |

        | Body | `password` | 是 | string | 密码，服务端加密后写入 sys_user.password_hash |


        主要数据表：sys_user、user_profile、user_login_session'
...
  /api/v1/agent-sessions:
    post:
      tags:
      - Agent智能体
      summary: 创建对话会话
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                agentId:
                  type: string
                  example: nebula
                  description: 智能体 ID，如 nebula/coach/sage
                title:
                  type: string
                  description: 标题
                  example: 二叉搜索树章节课后练习
                contextNodeId:
                  type: string
                  description: 上下文知识点 ID，对应 knowledge_node.id
                contextResourceId:
                  type: string
                  description: 上下文资源 ID，对应 learning_resource.id
                metadata:
                  type: object
                  additionalProperties: true
                  description: 扩展元数据
              required:
              - agentId
      responses:
        '200':
          $ref: '#/components/responses/Success'
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Body | `agentId` | 是 | string | 智能体 ID，如 nebula/coach/sage |

        | Body | `title` | 否 | string | 标题 |

        | Body | `contextNodeId` | 否 | string | 上下文知识点 ID，对应 knowledge_node.id |

        | Body | `contextResourceId` | 否 | string | 上下文资源 ID，对应 learning_resource.id |

        | Body | `metadata` | 否 | object | 扩展元数据 |


        主要数据表：agent_chat_session、agent_chat_message、agent_graph_thread、agent_memory、async_generation_task'
  /api/v1/agent-sessions/{sessionId}/messages:
    get:
      tags:
      - Agent智能体
      summary: 获取会话历史
      parameters:
      - $ref: '#/components/parameters/sessionId'
      - $ref: '#/components/parameters/page'
      - $ref: '#/components/parameters/pageSize'
      responses:
        '200':
          $ref: '#/components/responses/PageSuccess'
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Path | `sessionId` | 是 | string | Agent 会话 ID，对应 agent_chat_session.id |

        | Query | `page` | 否 | integer | 页码，从 1 开始 |

        | Query | `pageSize` | 否 | integer | 每页数量 |


        主要数据表：agent_chat_session、agent_chat_message、agent_graph_thread、agent_memory、async_generation_task'
    post:
      tags:
      - Agent智能体
      summary: 发送消息给智能体
      parameters:
      - $ref: '#/components/parameters/sessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                content:
                  type: string
                  description: 正文内容
                  example: 这里是正文内容
                contentType:
                  type: string
                  example: text
                  enum:
                  - text
                  - image
                  - file
                  - mixed
                  description: 内容类型
                attachments:
                  type: array
                  items:
                    type: object
                    additionalProperties: true
                  description: 附件列表，对应 sys_oss_file
                stream:
                  type: boolean
                  example: true
                  description: 是否使用流式响应
              required:
              - content
      responses:
        '200':
          $ref: '#/components/responses/Success'
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Path | `sessionId` | 是 | string | Agent 会话 ID，对应 agent_chat_session.id |

        | Body | `content` | 是 | string | 正文内容 |

        | Body | `contentType` | 否 | string | 内容类型 |

        | Body | `attachments` | 否 | array<object> | 附件列表，对应 sys_oss_file |

        | Body | `stream` | 否 | boolean | 是否使用流式响应 |


        主要数据表：agent_chat_session、agent_chat_message、agent_graph_thread、agent_memory、async_generation_task'
  /api/v1/agent-sessions/{sessionId}/stream:
    get:
      tags:
      - Agent智能体
      summary: Agent 流式响应
      parameters:
      - $ref: '#/components/parameters/sessionId'
      - $ref: '#/components/parameters/messageIdQuery'
      responses:
        '200':
          description: SSE 流式响应
          content:
            text/event-stream:
              schema:
                type: string
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Path | `sessionId` | 是 | string | Agent 会话 ID，对应 agent_chat_session.id |

        | Query | `messageId` | 是 | string | 消息 ID，对应 agent_chat_message.id 或 im_message.id |


        主要数据表：agent_chat_session、agent_chat_message、agent_graph_thread、agent_memory、async_generation_task'
...
  /api/v1/manage/auth/login:
    post:
      tags:
      - Manage管理端
      summary: 管理端登录
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                account:
                  type: string
                  description: 邮箱或用户名，对应 sys_user.email/sys_user.username
                  example: spark@example.com
                password:
                  type: string
                  description: 密码，服务端加密后写入 sys_user.password_hash
                  example: '12345678'
                captcha:
                  type: string
                  description: 验证码
              required:
              - account
              - password
      responses:
        '200':
          $ref: '#/components/responses/AuthSuccess'
      description: '参数设计：


        | 位置 | 参数名 | 必填 | 类型 | 说明 |

        | --- | --- | --- | --- | --- |

        | Body | `account` | 是 | string | 邮箱或用户名，对应 sys_user.email/sys_user.username |

        | Body | `password` | 是 | string | 密码，服务端加密后写入 sys_user.password_hash |

        | Body | `captcha` | 否 | string | 验证码 |


        主要数据表：sys_admin、sys_role、sys_permission、feedback_ticket、kb_document、sys_operation_log、sys_agent_prompt'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
      description: 页码，从 1 开始
    pageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        default: 10
        minimum: 1
        maximum: 100
      description: 每页数量
...
    sessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
      description: Agent 会话 ID，对应 agent_chat_session.id
    messageIdQuery:
      name: messageId
      in: query
      required: true
      schema:
        type: string
      description: 消息 ID，对应 agent_chat_message.id 或 im_message.id
...
  responses:
    Success:
      description: 成功
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
    PageSuccess:
      description: 分页成功
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiResponse'
            - type: object
              properties:
                data:
                  $ref: '#/components/schemas/PageResult'
    AuthSuccess:
      description: 登录或刷新成功
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiResponse'
            - type: object
              properties:
                data:
                  type: object
                  properties:
                    accessToken:
                      type: string
                    refreshToken:
                      type: string
                    expiresIn:
                      type: integer
                      example: 7200
  schemas:
    ApiResponse:
      type: object
      required:
      - code
      - message
      - data
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: success
        data:
          nullable: true
          oneOf:
          - type: object
            additionalProperties: true
          - type: array
            items: {}
          - type: string
          - type: number
          - type: boolean
        requestId:
          type: string
          example: req_20260505_xxxx
    PageResult:
      type: object
      properties:
        page:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 10
        total:
          type: integer
          example: 120
        list:
          type: array
          items:
            type: object
            additionalProperties: true